+++ pgmax/factors/logical.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 """Defines a logical factor"""
 
 import functools
@@ -148,42 +146,6 @@ class LogicalFactor(nodes.Factor):
             edge_states_offset=wirings[0].edge_states_offset,
         )
 
-    def compile_wiring(
-        self, vars_to_starts: Mapping[nodes.Variable, int]
-    ) -> LogicalWiring:
-        """Compile LogicalWiring for the LogicalFactor
-
-        Args:
-            vars_to_starts: A dictionary that maps variables to their global starting indices
-                For an n-state variable, a global start index of m means the global indices
-                of its n variable states are m, m + 1, ..., m + n - 1
-
-        Returns:
-             LogicalWiring for the LogicalFactor
-        """
-        var_states_for_edges = np.concatenate(
-            [
-                np.arange(variable.num_states) + vars_to_starts[variable]
-                for variable in self.variables
-            ]
-        )
-        num_parents = len(self.variables) - 1
-        relevant_state = (-self.edge_states_offset + 1) // 2
-        parents_edge_states = np.vstack(
-            [
-                np.zeros(num_parents, dtype=int),
-                np.arange(relevant_state, 2 * num_parents, 2, dtype=int),
-            ],
-        ).T
-        child_edge_state = np.array([2 * num_parents + relevant_state], dtype=int)
-        return LogicalWiring(
-            edges_num_states=self.edges_num_states,
-            var_states_for_edges=var_states_for_edges,
-            parents_edge_states=parents_edge_states,
-            children_edge_states=child_edge_state,
-            edge_states_offset=self.edge_states_offset,
-        )
-
 
 @dataclass(frozen=True, eq=False)
 class ORFactor(LogicalFactor):
@@ -217,175 +179,6 @@ class ANDFactor(LogicalFactor):
     """
 
     edge_states_offset: int = field(init=False, default=-1)
-
-
-import collections
-from typing import FrozenSet, OrderedDict, Type
-
-from pgmax.fg.groups import FactorGroup
-from pgmax.utils import cached_property
-
-
-@dataclass(frozen=True, eq=False)
-class LogicalFactorGroup(FactorGroup):
-    """Class to represent a group of LogicalFactors.
-
-    All factors in the group are assumed to have the same edge_states_offset.
-    Consequently, the factors are all ORFactors or ANDFactors.
-
-    Args:
-        edge_states_offset: Offset to go from a variable's relevant state to its other state
-            For ORFactors the edge_states_offset is 1, for ANDFactors the edge_states_offset is -1.
-    """
-
-    log_potentials: np.ndarray = field(init=False, default=np.empty((0,)))
-    edge_states_offset: int = field(init=False)
-
-    def __post_init__(self):
-        # TODO: move all asserts from EnumerationFactor to here
-        super().__post_init__()
-        pass
-
-    @cached_property
-    def factor_group_log_potentials_full(self) -> np.ndarray:
-        return self.log_potentials
-
-    def compile_wiring(self, vars_to_starts) -> LogicalWiring:
-        """Compile LogicalWiring for the LogicalFactorGroup
-
-        Args:
-            vars_to_starts: A dictionary that maps variables to their global starting indices
-                For an n-state variable, a global start index of m means the global indices
-                of its n variable states are m, m + 1, ..., m + n - 1
-
-        Returns:
-             LogicalWiring for the LogicalFactorGroup
-        """
-        # TODO: run in parallel?
-        import time
-
-        start = time.time()
-        relevant_state = (-self.edge_states_offset + 1) // 2
-
-        var_states_for_edges = []
-        for variable_and_num_states in self.variables_and_num_states:
-            variable, num_states = variable_and_num_states
-            this_var_states_for_edges = np.arange(
-                vars_to_starts[variable], vars_to_starts[variable] + num_states
-            )
-            var_states_for_edges.append(this_var_states_for_edges)
-        print(time.time() - start)
-
-        edges_num_states_cumsum = 0
-        parents_edge_states = []
-        children_edge_states = []
-        for factor_idx, variable_names_for_factor in enumerate(
-            self.variable_names_for_factors
-        ):
-            num_parents = len(variable_names_for_factor) - 1
-
-            # Note: edges_num_states_cumsum corresponds to the factor_to_msgs_start for the LogicalFactor
-            this_parents_edge_states = np.vstack(
-                [
-                    np.full(num_parents, fill_value=factor_idx, dtype=int),
-                    np.arange(
-                        edges_num_states_cumsum + relevant_state,
-                        edges_num_states_cumsum + 2 * num_parents,
-                        2,
-                        dtype=int,
-                    ),
-                ],
-            ).T
-            this_child_edge_state = (
-                edges_num_states_cumsum + 2 * num_parents + relevant_state
-            )
-
-            parents_edge_states.append(this_parents_edge_states)
-            children_edge_states.append(this_child_edge_state)
-            edges_num_states_cumsum += 2 * (num_parents + 1)
-        print(time.time() - start)
-
-        return LogicalWiring(
-            edges_num_states=self.factor_edges_num_states,
-            var_states_for_edges=np.concatenate(var_states_for_edges),
-            parents_edge_states=np.concatenate(parents_edge_states),
-            children_edge_states=np.array(children_edge_states),
-            edge_states_offset=self.edge_states_offset,
-        )
-
-
-@dataclass(frozen=True, eq=False)
-class ORFactorGroup(LogicalFactorGroup):
-    """Class to represent a group of ORFactors.
-
-    Args:
-        edge_states_offset: Offset to go from a variable's relevant state to its other state
-            For ORFactors the edge_states_offset is 1.
-    """
-
-    edge_states_offset: int = field(init=False, default=1)
-    factor_type: Type = field(init=False, default=ORFactor)
-
-    def _get_variables_to_factors(
-        self,
-    ) -> OrderedDict[FrozenSet, LogicalFactor]:
-        """Function that generates a dictionary mapping set of connected variables to factors.
-        This function is only called on demand when the user requires it.
-
-        Returns:
-            A dictionary mapping all possible set of connected variables to different factors.
-        """
-        variables_to_factors = collections.OrderedDict(
-            [
-                (
-                    frozenset(self.variable_names_for_factors[ii]),
-                    ORFactor(
-                        variables=tuple(
-                            self.variable_group[self.variable_names_for_factors[ii]]
-                        ),
-                    ),
-                )
-                for ii in range(len(self.variable_names_for_factors))
-            ]
-        )
-        return variables_to_factors
-
-
-@dataclass(frozen=True, eq=False)
-class ANDFactorGroup(LogicalFactorGroup):
-    """Class to represent a group of ANDFactors.
-
-    Args:
-        edge_states_offset: Offset to go from a variable's relevant state to its other state
-            For ANDFactors the edge_states_offset is -1.
-    """
-
-    edge_states_offset: int = field(init=False, default=-1)
-    factor_type: Type = field(init=False, default=ANDFactor)
-
-    def _get_variables_to_factors(
-        self,
-    ) -> OrderedDict[FrozenSet, LogicalFactor]:
-        """Function that generates a dictionary mapping set of connected variables to factors.
-        This function is only called on demand when the user requires it.
-
-        Returns:
-            A dictionary mapping all possible set of connected variables to different factors.
-        """
-        variables_to_factors = collections.OrderedDict(
-            [
-                (
-                    frozenset(self.variable_names_for_factors[ii]),
-                    ANDFactor(
-                        variables=tuple(
-                            self.variable_group[self.variable_names_for_factors[ii]]
-                        ),
-                    ),
-                )
-                for ii in range(len(self.variable_names_for_factors))
-            ]
-        )
-        return variables_to_factors
 
 
 @functools.partial(jax.jit, static_argnames=("temperature"))
